+++ common/src/jsMain/kotlin/dev/inmo/micro_utils/common/SelectFile.kt
@@ -7,13 +7,13 @@ import org.w3c.dom.HTMLInputElement
 import org.w3c.files.get
 
 fun selectFile(
-    inputSetup: HTMLInputElement.() -> Unit = {},
+    inputSetup: (HTMLInputElement) -> Unit = {},
     onFailure: (Throwable) -> Unit = {},
     onFile: (MPPFile) -> Unit
 ) {
     (document.createElement("input") {
         (this as HTMLInputElement).apply {
-            inputSetup()
+            inputSetup(this)
             onchange = {
                 runCatching {
                     files ?.get(0) ?: error("File must not be null")